# backend/app/api/static/docs/openapi-3.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Track, Manage, and Organize your Things.",
        "title": "Homebox API",
        "contact": {
            "name": "Homebox Team",
            "url": "https://discord.homebox.software"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/users/forgot-password": {
            "post": {
                "description": "Sends a password reset email if the address is associated with a local account.\nAlways returns 204 on success to avoid leaking whether the email is registered.",
                "tags": [
                    "Authentication"
                ],
                "summary": "Request Password Reset",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/v1.ForgotPasswordRequest"
                            }
                        }
                    },
                    "description": "Email",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "missing or invalid request body, or empty email field",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "demo mode is enabled or local login is disabled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/v1/users/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "auth provider",
                        "name": "provider",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/v1.LoginForm"
                            }
                        },
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/v1.LoginForm"
                            }
                        }
                    },
                    "description": "Login Data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v1.TokenResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/users/login/oidc/callback": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "OIDC Callback Handler",
                "parameters": [
                    {
                        "description": "Authorization code",
                        "name": "code",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "State parameter",
                        "name": "state",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "302": {
                        "description": "Found"
                    }
                }
            }
        },
        "/v1/users/logout": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User Logout",
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/users/refresh": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "handleAuthRefresh returns a handler that will issue a new token from an existing token.\nThis does not validate that the user still exists within the database.",
                "tags": [
                    "Authentication"
                ],
                "summary": "User Token Refresh",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/v1/users/reset-password": {
            "post": {
                "description": "Consumes a single-use reset token and changes the user's password.\nOn success, all existing sessions for the user are revoked.",
                "tags": [
                    "Authentication"
                ],
                "summary": "Reset Password",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/v1.ResetPasswordRequest"
                            }
                        }
                    },
                    "description": "Token + new password",
                    "required": true
                },
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "invalid request body, password shorter than the minimum length, or token that is invalid / expired / already used",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "demo mode is enabled or local login is disabled",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "internal error while processing the request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "components": {
        "securitySchemes": {
            "Bearer": {
                "description": "\"Type 'Bearer TOKEN' to correctly set the API Key\"",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "v1.ForgotPasswordRequest": {
                "type": "object",
                "required": [
                    "email"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@example.com"
                    }
                }
            },
            "v1.LoginForm": {
                "type": "object",
                "properties": {
                    "password": {
                        "type": "string",
                        "example": "admin"
                    },
                    "stayLoggedIn": {
                        "type": "boolean"
                    },
                    "username": {
                        "type": "string",
                        "example": "admin@admin.com"
                    }
                }
            },
...
            "v1.ResetPasswordRequest": {
                "type": "object",
                "required": [
                    "password",
                    "token"
                ],
                "properties": {
                    "password": {
                        "type": "string",
                        "minLength": 6
                    },
                    "token": {
                        "type": "string",
                        "minLength": 20
                    }
                }
            },
...
            "v1.TokenResponse": {
                "type": "object",
                "properties": {
                    "attachmentToken": {
                        "type": "string"
                    },
                    "expiresAt": {
                        "type": "string"
                    },
                    "token": {
                        "type": "string"
                    }
                }
            },
...
        }
    }
}